pyproject.toml
[tool.poetry]
name = "shop-api"
version = "0.1.0"
description = ""
authors = ["makers"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.15.1"
psycopg2-binary = "^2.9.9"
python-decouple = "^3.8"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.3.0"
django-filter = "^24.2"
drf-yasg = "^1.21.7"
python-slugify = "^8.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
